+++ carioca-report/report-android/src/main/kotlin/com/rubensousa/carioca/report/android/InstrumentedReportRule.kt
@@ -167,7 +167,12 @@ open class InstrumentedReportRule internal constructor(
         action: InstrumentedBlockingTest.() -> Unit,
     ) {
         val currentTest = getCurrentTest<InstrumentedBlockingTest>()
-        action(currentTest)
+        try {
+            action(currentTest)
+        } catch (error: Throwable) {
+            currentTest.onFailed(error)
+            throw error
+        }
     }
 
 }