+++ carioca-report/report-android/src/main/kotlin/com/rubensousa/carioca/report/android/stage/InstrumentedTestReport.kt
@@ -26,7 +26,9 @@ import com.rubensousa.carioca.report.android.recording.ReportRecording
 import com.rubensousa.carioca.report.android.recording.ScreenRecorder
 import com.rubensousa.carioca.report.android.screenshot.ScreenshotDelegate
 import com.rubensousa.carioca.report.android.storage.FileIdGenerator
+import com.rubensousa.carioca.report.android.storage.ReportStorageProvider
 import com.rubensousa.carioca.report.runtime.ReportProperty
+import com.rubensousa.carioca.report.runtime.ReportStatus
 import com.rubensousa.carioca.report.runtime.StageAttachment
 import com.rubensousa.carioca.report.runtime.StageStack
 import com.rubensousa.carioca.report.runtime.TestMetadata
@@ -40,7 +42,7 @@ import com.rubensousa.carioca.report.runtime.TestMetadata
  */
 abstract class InstrumentedTestReport internal constructor(
     outputPath: String,
-    storageProvider: com.rubensousa.carioca.report.android.storage.ReportStorageProvider,
+    storageProvider: ReportStorageProvider,
     val metadata: TestMetadata,
     protected val interceptors: List<CariocaInstrumentedInterceptor>,
     private val recordingOptions: RecordingOptions,
@@ -60,7 +62,7 @@ abstract class InstrumentedTestReport internal constructor(
         screenshotDelegate: ScreenshotDelegate,
         reporter: InstrumentedReporter,
         interceptors: List<CariocaInstrumentedInterceptor>,
-        storageProvider: com.rubensousa.carioca.report.android.storage.ReportStorageProvider,
+        storageProvider: ReportStorageProvider,
     ) : this(
         outputPath = outputPath,
         metadata = metadata,
@@ -107,6 +109,15 @@ abstract class InstrumentedTestReport internal constructor(
     }
 
     fun onFailed(error: Throwable) {
+        /**
+         * We also call [onFailed] internally inside the test action reports,
+         * to ensure the report file is saved before the instrumentation marks the test as over.
+         * Since tests can just use the rule without using the stage reports,
+         * we can't distinguish the 2 use-cases without doing this
+         */
+        if (getExecutionMetadata().status == ReportStatus.FAILED) {
+            return
+        }
         // Take a screenshot asap to record the state on failures
         screenshotDelegate.takeScreenshot(this, "Screenshot of failure")