+++ docs/changelog.md
@@ -2,6 +2,6 @@
 
 Libraries:
 
-- [Android Allure Gradle Plugin](changelog_android_allure.md)
-- [Test reports](changelog_reports.md)
-- [JUnit4 Rules](changelog_junit4_rules.md)
+- [Android Allure Gradle Plugin](changelog-android-allure)
+- [Test reports](changelog-reports)
+- [JUnit4 Rules](changelog-junit4-rules)